+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add: Spec Ports DB dialog now allows to edit entire DB as text.
 - Add: Read Only mode can be toggled in a file. See new button near "Rename".
 - Add: Display Hex and Char values for hovered tile.
+- Add: Reordering levels in file. See "Levels order" in levels buttons.
 - Fix: Some UI texts correction in progress. Thanks to Lana Snik-Snak!
 - Fix: H/V/X Ports when set as Special wasn't marked Special in "Copy as Image".
 - Fix: Copy/Export Level/Selection icons and compact layout.